feat: add setuptools-rust backend with bracket-notation auto-detection
Add a setuptools_rust_build backend that shares Rust toolchain setup
with the existing maturin backend via extracted common modules:
- rust_common.py: shared Python-side Rust env configuration
- rust_common_sdist_hook.bzl: shared cargo crate vendoring
- cargo_lock_repo.bzl: shared repo rule for cargo lock targets

Introduce bracket notation in pyproject_backends for automatic backend
detection based on build-system.requires. For example, registering
'setuptools.build_meta[setuptools-rust]' causes packages that declare
setuptools.build_meta as their backend AND list setuptools-rust in their
requires to automatically use setuptools_rust_build. The most specific
match wins (most required packages satisfied).

This replaces the earlier 'supersedes' mechanism with a declarative
approach that requires no per-package overrides for basic detection.
