+++ modules/backend_maturin/private/BUILD.bazel
@@ -13,7 +13,24 @@ bzl_library(
 bzl_library(
     name = "maturin_sdist_hook",
     srcs = ["maturin_sdist_hook.bzl"],
+    deps = [":rust_common_sdist_hook"],
+)
+
+bzl_library(
+    name = "setuptools_rust_sdist_hook",
+    srcs = ["setuptools_rust_sdist_hook.bzl"],
+    deps = [":rust_common_sdist_hook"],
+)
+
+bzl_library(
+    name = "rust_common_sdist_hook",
+    srcs = ["rust_common_sdist_hook.bzl"],
     deps = [":cargo"],
+)
+
+bzl_library(
+    name = "cargo_lock_repo",
+    srcs = ["cargo_lock_repo.bzl"],
 )
 
 bzl_library(
