+++ modules/backend_maturin/private/tools/BUILD.bazel
@@ -1,0 +1,46 @@
+load("@rules_python//python:defs.bzl", "py_binary", "py_library")
+
+package(default_visibility = ["//visibility:public"])
+
+py_library(
+    name = "generate_cargo_lock_lib",
+    srcs = ["generate_cargo_lock.py"],
+    imports = ["../.."],
+)
+
+py_binary(
+    name = "generate_cargo_lock",
+    srcs = ["generate_cargo_lock.py"],
+    deps = [":generate_cargo_lock_lib"],
+)
+
+py_library(
+    name = "rust_common_lib",
+    srcs = ["rust_common.py"],
+    imports = ["../.."],
+    deps = [
+        "@rules_pycross//pycross/private/build/tools:utils",
+    ],
+)
+
+py_binary(
+    name = "maturin_builder",
+    srcs = ["maturin_builder.py"],
+    imports = ["../.."],
+    deps = [
+        ":rust_common_lib",
+        "@rules_pycross//pycross/private/build/tools:utils",
+        "@rules_pycross//pycross/private/build/tools/crossenv",
+    ],
+)
+
+py_binary(
+    name = "setuptools_rust_builder",
+    srcs = ["setuptools_rust_builder.py"],
+    imports = ["../.."],
+    deps = [
+        ":rust_common_lib",
+        "@rules_pycross//pycross/private/build/tools:utils",
+        "@rules_pycross//pycross/private/build/tools/crossenv",
+    ],
+)
