--- modules/backend_maturin/tools/BUILD.bazel
@@ -1,15 +1,0 @@
-load("@rules_python//python:defs.bzl", "py_binary", "py_library")
-
-package(default_visibility = ["//visibility:public"])
-
-py_library(
-    name = "generate_cargo_lock_lib",
-    srcs = ["generate_cargo_lock.py"],
-    imports = [".."],
-)
-
-py_binary(
-    name = "generate_cargo_lock",
-    srcs = ["generate_cargo_lock.py"],
-    deps = [":generate_cargo_lock_lib"],
-)
