+++ pycross/backend.bzl
@@ -9,7 +9,6 @@ load(
     "//pycross/private:lock_attrs.bzl",
     _BUILD_SYSTEM_ATTRS = "BUILD_SYSTEM_ATTRS",
     _CC_BUILD_SYSTEM_ATTRS = "CC_BUILD_SYSTEM_ATTRS",
-    _MATURIN_OVERRIDE_ATTRS = "MATURIN_OVERRIDE_ATTRS",
 )
 load(
     "//pycross/private:override_helpers.bzl",
@@ -96,4 +95,3 @@ create_overrides_repo = _create_overrides_repo
 encode_build_system_attrs = _encode_build_system_attrs
 BUILD_SYSTEM_ATTRS = _BUILD_SYSTEM_ATTRS
 CC_BUILD_SYSTEM_ATTRS = _CC_BUILD_SYSTEM_ATTRS
-MATURIN_OVERRIDE_ATTRS = _MATURIN_OVERRIDE_ATTRS
