+++ tests/e2e/run_integration_test.sh
@@ -8,6 +8,7 @@ WORKSPACE_ROOT=$(dirname $(dirname "$E2E_DIR"))
 
 WORKSPACE_DIR="${WORKSPACE_ROOT}/$1"
 echo "Running Bazel integration test in ${WORKSPACE_DIR}..."
+echo "DEBUG: RULES_PYCROSS_DEBUG=${RULES_PYCROSS_DEBUG:-}"
 
 if [[ ! -d "${WORKSPACE_DIR}" ]]; then
   echo "Error: Directory ${WORKSPACE_DIR} does not exist."
@@ -32,12 +33,18 @@ if [[ -x "run.sh" ]]; then
   echo "Running custom run.sh..."
   ./run.sh
 else
+  BAZEL_ARGS=()
+  if [ -n "${RULES_PYCROSS_DEBUG:-}" ]; then
+    BAZEL_ARGS+=("--action_env=RULES_PYCROSS_DEBUG=1")
+    BAZEL_ARGS+=("--remote_accept_cached=false")
+    BAZEL_ARGS+=("--nouse_action_cache")
+  fi
   echo "Running bazel build //..."
-  bazel build //...
+  bazel build "${BAZEL_ARGS[@]}" //...
   # Query if there are any test targets in the workspace
   if bazel query "tests(//...)" --output=label 2>/dev/null | grep -q .; then
     echo "Running bazel test //..."
-    bazel test //...
+    bazel test "${BAZEL_ARGS[@]}" //...
   else
     echo "No test targets found, skipping bazel test."
   fi
